+++ CHANGELOG.md
@@ -16,8 +16,13 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 - `prevent-canvas` scriptlet [#451]
 
+### Changed
+
+- improve `prevent-xhr` — modify response [#415]
+
 [Unreleased]: https://github.com/AdguardTeam/Scriptlets/compare/v1.12.1...HEAD
 [#451]: https://github.com/AdguardTeam/Scriptlets/issues/451
+[#415]: https://github.com/AdguardTeam/Scriptlets/issues/415
 
 ## [v1.12.1] - 2024-09-20